requirements.txt
asgiref==3.4.1
attrs==22.2.0
autobahn==21.2.1
Automat==22.10.0
certifi==2024.2.2
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.25
django-allauth==0.52.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2024.1
requests==2.27.1
requests-oauthlib==2.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==21.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.18
zope.interface==5.5.2
